views refactored, permissions added to all apis

+++ healthprofile/views.py
@@ -3,9 +3,19 @@ from healthprofile.serializers import (HealthProfileSerializer,
 from healthprofile.models import HealthProfile, Illness
 from rest_framework.views import APIView
 from rest_framework.response import Response
+from rest_framework import permissions, generics
 
 
-class HealthProfileView(APIView):
+class IsUserOwnerHealthProfile(permissions.BasePermission):
+    def has_object_permission(self, request, view, health_profile_obj):
+        return (health_profile_obj.user_id.id == request.user.id)
+
+
+class HealthProfileView(generics.GenericAPIView):
+    permission_classes = [
+        IsUserOwnerHealthProfile
+    ]
+
     # to get Logged in User's Health profile
     def get(self, request, format=None):
         user = request.user
@@ -50,6 +60,7 @@ class HealthProfileView(APIView):
     # edit health profile
     def patch(self, request, pk):
         object = self.get_object(pk)
+        self.check_object_permissions(self.request, object)
         serializer = HealthProfileSerializer(
             object, data=request.data, partial=True)
         if serializer.is_valid():
@@ -66,7 +77,17 @@ class HealthProfileView(APIView):
         )
 
 
+class IsUserOwnerIllnessProfile(permissions.BasePermission):
+    def has_object_permission(self, request, view, illness_profile_obj):
+        return (illness_profile_obj.medical_profile_id
+                .user_id.id == request.user.id)
+
+
 class IllnessView(APIView):
+    permission_classes = [
+       IsUserOwnerIllnessProfile
+    ]
+
     # add a new illness
     def post(self, request, *args, **kwargs):
         context = {'request': request}
@@ -74,6 +95,7 @@ class IllnessView(APIView):
         if serializer.is_valid():
             data = serializer.save()
             profile = HealthProfile.objects.get(pk=data.medical_profile_id.id)
+            self.check_object_permissions(self.request, profile)
             illness = profile.illness.all() or []
             profile = HealthProfileSerializer(profile, context=request).data
             profile['illnesses'] = IllnessSerializer(
@@ -92,6 +114,7 @@ class IllnessView(APIView):
     # edit an illness object
     def patch(self, request, pk):
         object = self.get_object(pk)
+        self.check_object_permissions(self.request, object)
         serializer = IllnessSerializer(object, data=request.data, partial=True)
         if serializer.is_valid():
             data = serializer.save()
@@ -111,6 +134,7 @@ class IllnessView(APIView):
 
     def delete(self, request, pk):
         obj = Illness.objects.get(pk=pk)
+        self.check_object_permissions(self.request, object)
         profile = HealthProfile.objects.get(pk=obj.medical_profile_id.id)
         obj.delete()
         illness = profile.illness.all() or []
